requirements.txt
spidev
smbus
gpiozero
Pillow
lgpio
gps3
